+++ templates/modals.html
@@ -295,7 +295,7 @@
     </div>
 </div>
 
-{% include 'modals/errors.html' %}
+{% include 'modals/action_list.html' %}
 
 {% include 'modals/settings.html' %}
 {% include 'modals/shortcuts.html' %}
